Реализована проверка сравнением двух параметров
