+++ src/RevitKrChecker/Models/Check/ParamCheck.cs
@@ -1,15 +1,17 @@
+// Ignore Spelling: Tooltip
+
 using System;
 using System.Collections.Generic;
 using System.Linq;
 
 using Autodesk.Revit.DB;
 
-using dosymep.Revit;
-
 using RevitKrChecker.Models.Rule;
+using RevitKrChecker.Models.Services;
 
 namespace RevitKrChecker.Models.Check {
     public class ParamCheck : ICheck {
+        private readonly ParamService _paramService;
 
         public ParamCheck(string checkName,
                           string targetParamName,
@@ -21,6 +23,8 @@ namespace RevitKrChecker.Models.Check {
             TargetParamLevel = targetParamLevel;
             CheckRule = checkRule ?? throw new ArgumentNullException(nameof(checkRule));
             TrueValues = trueValues ?? throw new ArgumentNullException(nameof(trueValues));
+
+            _paramService = new ParamService();
         }
 
         public string CheckName { get; }
@@ -28,29 +32,6 @@ namespace RevitKrChecker.Models.Check {
         public ParamLevel TargetParamLevel { get; }
         public ICheckRule CheckRule { get; }
         public List<string> TrueValues { get; }
-
-
-        private List<Parameter> GetParamsToCheck(Element element, string paramName, ParamLevel paramLevel) {
-            Document doc = element.Document;
-            switch(paramLevel) {
-                case ParamLevel.Instance:
-                    return new List<Parameter> {
-                        element.GetParam(paramName)
-                    };
-                case ParamLevel.Type:
-                    return new List<Parameter> {
-                        doc.GetElement(element.GetTypeId()).GetParam(paramName)
-                    };
-                case ParamLevel.Material:
-                    return element
-                        .GetMaterialIds(false)
-                        .Select(id => doc.GetElement(id))
-                        .Select(material => material.GetParam(paramName))
-                        .ToList();
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(paramLevel), "Задан неизвестный уровень параметра");
-            }
-        }
 
         private bool CheckAnyTrueValues(string targetParamValue) {
             return TrueValues.Any(trueValue => CheckRule.Check(targetParamValue, trueValue));
@@ -60,7 +41,7 @@ namespace RevitKrChecker.Models.Check {
             if(element == null)
                 throw new ArgumentNullException(nameof(element));
 
-            var targetParams = GetParamsToCheck(element, TargetParamName, TargetParamLevel);
+            var targetParams = _paramService.GetParamsToCheck(element, TargetParamName, TargetParamLevel);
 
             foreach(var targetParam in targetParams) {
                 string targetParamValue = targetParam.AsValueString();
@@ -70,7 +51,6 @@ namespace RevitKrChecker.Models.Check {
                     return false;
                 }
             }
-
             info = null;
             return true;
         }
